+++ motor.h
@@ -38,7 +38,7 @@ extern struct Bldc_phase_state bldc_phase_state;
 extern unsigned int motor_speed_erps;
 extern unsigned int PWM_cycles_counter;
 extern int motor_rotor_position;
-extern int position_correction_value;
+extern float position_correction_value;
 extern int duty_cycle;
 
 extern unsigned int adc_phase_a_current_offset;
